feat(state-management): this commit might be very simple but, it shows how a Store Action dispatch changes values at for all variables which subscribes to that Selector
