+++ src/app/app.component.html
@@ -16,15 +16,6 @@
             <p-panelMenu [model]="items" [style]="{'width':'100%', 'height': '100vh'}"></p-panelMenu>
           </ng-template> -->
           <ng-template pTemplate>
-            <!-- <div class="d-flex">
-              <p>{{currentTemplateLabel}}</p>
-              &nbsp;
-              <p>{{currentTemplateNumber}}</p>
-              <br>
-              <p>
-                <button class="btn btn-default" type="button" (click)="doThis()">Test Here!</button>
-              </p>
-            </div> -->
             <!-- Panel 2 -->
             <div #customcontainer1 class="custom-container-1 d-flex" [style]="{ 'padding-top': '.75rem', 'padding-bottom': '.75rem' }">
               <!-- <ng-container *ngFor="let i of [1, 2, 3, 4, 5] | slice: 0  : containerWidth" [ngTemplateOutlet]="Cards"> -->
